moving openssl tls server/client enclave initialization functions to utility file
