+++ tests/tls_e2e/common/openssl_utility.cpp
@@ -244,3 +244,54 @@ int write_to_session_peer(
 exit:
     return ret;
 }
+
+int initalize_ssl_context(SSL_CTX*& ctx)
+{
+    int ret = -1;
+    if ((ctx = SSL_CTX_new(SSLv23_server_method())) == NULL)
+    {
+        OE_TRACE_ERROR(TLS_SERVER " unable to create a new SSL context\n");
+        goto exit;
+    }
+    // choose TLSv1.2 by excluding SSLv2, SSLv3 ,TLS 1.0 and TLS 1.1
+    SSL_CTX_set_options(ctx, SSL_OP_NO_SSLv2);
+    SSL_CTX_set_options(ctx, SSL_OP_NO_SSLv3);
+    SSL_CTX_set_options(ctx, SSL_OP_NO_TLSv1);
+    SSL_CTX_set_options(ctx, SSL_OP_NO_TLSv1_1);
+    SSL_CTX_set_verify(ctx, SSL_VERIFY_PEER, &cert_verify_callback);
+    ret = 0;
+exit:
+    return ret;
+}
+
+int load_ssl_certificates_and_keys(SSL_CTX* ctx, X509*& cert, EVP_PKEY*& pkey)
+{
+    int ret = -1;
+    if (generate_certificate_and_pkey(cert, pkey) != OE_OK)
+    {
+        OE_TRACE_ERROR(TLS_SERVER "Cannot generate certificate and pkey\n");
+        goto exit;
+    }
+    if (!SSL_CTX_use_certificate(ctx, cert))
+    {
+        OE_TRACE_ERROR(TLS_SERVER "Cannot load certificate on the server\n");
+        goto exit;
+    }
+
+    if (!SSL_CTX_use_PrivateKey(ctx, pkey))
+    {
+        OE_TRACE_ERROR(TLS_SERVER "Cannot load private key on the server\n");
+        goto exit;
+    }
+
+    /* verify private key */
+    if (!SSL_CTX_check_private_key(ctx))
+    {
+        OE_TRACE_ERROR(TLS_SERVER
+                       "Private key does not match the public certificate\n");
+        goto exit;
+    }
+    ret = 0;
+exit:
+    return ret;
+}
